LPS-83319 Moves the code only used when editing
